requirements.txt
certifi==2020.12.5
chardet==4.0.0
Django==2.2.17
django-filter==2.4.0
django-rest-auth==0.9.5
djangorestframework==3.12.2
djangorestframework-api-key==2.0.0
djangorestframework-jwt==1.11.0
drf-nested-routers==0.92.5
idna==2.10
Markdown==3.3.3
PyJWT==1.7.1
pytz==2020.4
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
